requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
filelock==3.12.0
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.2
openai==0.27.2
Pillow==9.4.0
pycparser==2.21
python-dotenv==1.0.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sounddevice==0.4.6
soundfile==0.12.1
srt==3.5.2
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.15
vosk==0.3.45
websockets==10.4
yarl==1.8.2
